refactor: extract ical checks into own module
Starts to split the monolithic check modules for wakeups and activity
checks into modules organized by function. Often, there is some overlap
between activity and wakeup checks used for similar purposes. Therefore,
we can assume that they will change in parallel. Moreover, also the
utility functions are related to realizing the ical functionality.
Therefore, they are also moved here from their own module. Hence, this
refactoring ensures the common closure principle on the module level.
