+++ tests/test_checks_ical/all-day-recurring-exclusions.ics
