+++ tests/test_checks_ical/simple-recurring.ics
