--- tests/test_checks_wakeup/old-event.ics
@@ -1,32 +1,0 @@
-BEGIN:VCALENDAR
-PRODID:-//Mozilla.org/NONSGML Mozilla Calendar V1.1//EN
-VERSION:2.0
-BEGIN:VTIMEZONE
-TZID:Europe/Berlin
-BEGIN:DAYLIGHT
-TZOFFSETFROM:+0100
-TZOFFSETTO:+0200
-TZNAME:CEST
-DTSTART:19700329T020000
-RRULE:FREQ=YEARLY;BYDAY=-1SU;BYMONTH=3
-END:DAYLIGHT
-BEGIN:STANDARD
-TZOFFSETFROM:+0200
-TZOFFSETTO:+0100
-TZNAME:CET
-DTSTART:19701025T030000
-RRULE:FREQ=YEARLY;BYDAY=-1SU;BYMONTH=10
-END:STANDARD
-END:VTIMEZONE
-BEGIN:VEVENT
-CREATED:20180603T194125Z
-LAST-MODIFIED:20180603T194144Z
-DTSTAMP:20180603T194144Z
-UID:6ff13ee1-e548-41b1-8e08-d7725423743a
-SUMMARY:long-event
-DTSTART;TZID=Europe/Berlin:20040605T130000
-DTEND;TZID=Europe/Berlin:20040605T150000
-TRANSP:OPAQUE
-SEQUENCE:1
-END:VEVENT
-END:VCALENDAR
